+++ src/components/custom-emojis-modal.jsx
@@ -300,6 +300,7 @@ function CustomEmojisModal({
             autocapitalize="off"
             spellCheck="false"
             dir="auto"
+            enterKeyHint="search"
             defaultValue={defaultSearchTerm || ''}
           />
         </form>