+++ src/components/poll.jsx
@@ -191,6 +191,89 @@ export default function Poll({
         </form>
       )}
       <p class="poll-meta">
+        <span class="spacer">
+          <span class="ib">
+            <Plural
+              value={votesCount}
+              one={
+                <Trans>
+                  <span title={votesCount}>{shortenNumber(votesCount)}</span>{' '}
+                  vote
+                </Trans>
+              }
+              other={
+                <Trans>
+                  <span title={votesCount}>{shortenNumber(votesCount)}</span>{' '}
+                  votes
+                </Trans>
+              }
+            />
+          </span>
+          {!!votersCount && votersCount !== votesCount && (
+            <>
+              {' '}
+              &bull;{' '}
+              <span class="ib">
+                <Plural
+                  value={votersCount}
+                  one={
+                    <Trans>
+                      <span title={votersCount}>
+                        {shortenNumber(votersCount)}
+                      </span>{' '}
+                      voter
+                    </Trans>
+                  }
+                  other={
+                    <Trans>
+                      <span title={votersCount}>
+                        {shortenNumber(votersCount)}
+                      </span>{' '}
+                      voters
+                    </Trans>
+                  }
+                />
+              </span>
+            </>
+          )}{' '}
+          &bull;{' '}
+          {expired ? (
+            !!expiresAtDate ? (
+              <span class="ib">
+                <Trans>
+                  Ended <RelativeTime datetime={expiresAtDate} />
+                </Trans>
+              </span>
+            ) : (
+              t`Ended`
+            )
+          ) : !!expiresAtDate ? (
+            <span class="ib">
+              <Trans>
+                Ending <RelativeTime datetime={expiresAtDate} />
+              </Trans>
+            </span>
+          ) : (
+            t`Ending`
+          )}
+        </span>
+        {!voted && !expired && !readOnly && optionsHaveVoteCounts && (
+          <button
+            type="button"
+            class="plain small"
+            disabled={uiState === 'loading'}
+            onClick={(e) => {
+              e.preventDefault();
+              setShowResults(!showResults);
+            }}
+            title={showResults ? t`Hide results` : t`Show results`}
+          >
+            <Icon
+              icon={showResults ? 'eye-open' : 'eye-close'}
+              alt={showResults ? t`Hide results` : t`Show results`}
+            />{' '}
+          </button>
+        )}
         {!expired && !readOnly && (
           <button
             type="button"
@@ -209,74 +292,6 @@ export default function Poll({
           >
             <Icon icon="refresh" alt={t`Refresh`} />
           </button>
-        )}
-        {!voted && !expired && !readOnly && optionsHaveVoteCounts && (
-          <button
-            type="button"
-            class="plain small"
-            disabled={uiState === 'loading'}
-            onClick={(e) => {
-              e.preventDefault();
-              setShowResults(!showResults);
-            }}
-            title={showResults ? t`Hide results` : t`Show results`}
-          >
-            <Icon
-              icon={showResults ? 'eye-open' : 'eye-close'}
-              alt={showResults ? t`Hide results` : t`Show results`}
-            />{' '}
-          </button>
-        )}
-        {!expired && !readOnly && ' '}
-        <Plural
-          value={votesCount}
-          one={
-            <Trans>
-              <span title={votesCount}>{shortenNumber(votesCount)}</span> vote
-            </Trans>
-          }
-          other={
-            <Trans>
-              <span title={votesCount}>{shortenNumber(votesCount)}</span> votes
-            </Trans>
-          }
-        />
-        {!!votersCount && votersCount !== votesCount && (
-          <>
-            {' '}
-            &bull;{' '}
-            <Plural
-              value={votersCount}
-              one={
-                <Trans>
-                  <span title={votersCount}>{shortenNumber(votersCount)}</span>{' '}
-                  voter
-                </Trans>
-              }
-              other={
-                <Trans>
-                  <span title={votersCount}>{shortenNumber(votersCount)}</span>{' '}
-                  voters
-                </Trans>
-              }
-            />
-          </>
-        )}{' '}
-        &bull;{' '}
-        {expired ? (
-          !!expiresAtDate ? (
-            <Trans>
-              Ended <RelativeTime datetime={expiresAtDate} />
-            </Trans>
-          ) : (
-            t`Ended`
-          )
-        ) : !!expiresAtDate ? (
-          <Trans>
-            Ending <RelativeTime datetime={expiresAtDate} />
-          </Trans>
-        ) : (
-          t`Ending`
         )}
       </p>
     </div>