+++ src/components/related-actions.jsx
@@ -13,6 +13,7 @@ import states from '../utils/states';
 import { getCurrentAccountID, updateAccount } from '../utils/store-utils';
 import supports from '../utils/supports';
 
+import { handleScannerClick } from './account-info';
 import AddRemoveListsSheet from './add-remove-lists-sheet';
 import Icon from './icon';
 import Loader from './loader';
@@ -227,6 +228,23 @@ function RelatedActions({
               <span>{privateNote}</span>
             </button>
           )}
+          {currentAuthenticated && isSelf && (
+            <button
+              type="button"
+              class="plain"
+              onClick={() => {
+                states.showQrCodeModal = {
+                  text: url,
+                  arena: avatarStatic,
+                  backgroundMask: headerStatic,
+                  caption: acct.includes('@') ? acct : `${acct}@${instance}`,
+                  onScannerClick: handleScannerClick,
+                };
+              }}
+            >
+              <Icon icon="qrcode" alt={t`QR code`} />
+            </button>
+          )}
           <Menu2
             portal={{
               target: document.body,
@@ -241,8 +259,8 @@ function RelatedActions({
             position="anchor"
             overflow="auto"
             menuButton={
-              <button type="button" class="plain" disabled={loading}>
-                <Icon icon="more" size="l" alt={t`More`} />
+              <button type="button" class="plain4" disabled={loading}>
+                <Icon icon="more2" size="l" alt={t`More`} />
               </button>
             }
             onMenuChange={(e) => {
@@ -459,7 +477,7 @@ function RelatedActions({
                 <>
                   <MenuItem onClick={() => setRenderEndorsements(true)}>
                     <Icon icon="endorsement" />
-                    Show featured profiles
+                    <Trans>Show featured profiles</Trans>
                   </MenuItem>
                   <MenuDivider />
                 </>
@@ -538,10 +556,11 @@ function RelatedActions({
                   arena: avatarStatic,
                   backgroundMask: headerStatic,
                   caption: acct.includes('@') ? acct : `${acct}@${instance}`,
+                  onScannerClick: handleScannerClick,
                 };
               }}
             >
-              <Icon icon="qrcode-2-line" />
+              <Icon icon="qrcode" />
               <span>
                 <Trans>QR code</Trans>
               </span>