+++ src/utils/states.js
@@ -57,6 +57,8 @@ const states = proxy({
   showEmbedModal: false,
   showReportModal: false,
   showQrCodeModal: false,
+  showQrScannerModal: false,
+  showImportExportAccounts: false,
   // Shortcuts
   shortcuts: [],
   // Settings
@@ -73,7 +75,6 @@ const states = proxy({
     mediaAltGenerator: false,
     composerGIFPicker: false,
     cloakMode: false,
-    groupedNotificationsAlpha: false,
   },
 });
 
@@ -108,8 +109,6 @@ export function initStates() {
   states.settings.composerGIFPicker =
     store.account.get('settings-composerGIFPicker') ?? false;
   states.settings.cloakMode = store.account.get('settings-cloakMode') ?? false;
-  states.settings.groupedNotificationsAlpha =
-    store.account.get('settings-groupedNotificationsAlpha') ?? false;
 }
 
 subscribeKey(states, 'notificationsLast', (v) => {
@@ -159,9 +158,6 @@ subscribe(states, (changes) => {
     if (path.join('.') === 'settings.cloakMode') {
       store.account.set('settings-cloakMode', !!value);
     }
-    if (path.join('.') === 'settings.groupedNotificationsAlpha') {
-      store.account.set('settings-groupedNotificationsAlpha', !!value);
-    }
   }
 });
 
@@ -179,6 +175,8 @@ export function hideAllModals() {
   states.showEmbedModal = false;
   states.showReportModal = false;
   states.showQrCodeModal = false;
+  states.showQrScannerModal = false;
+  states.showImportExportAccounts = false;
 }
 
 export function statusKey(id, instance) {